requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
jplephem==2.17
matplotlib==3.5.1
numpy==1.22.2
openpyxl==3.0.9
psycopg2==2.9.3
pymemcache==3.5.1
pytz==2021.3
requests==2.27.1
sgp4==2.20
skyfield==1.42
SQLAlchemy==1.4.31
urllib3==1.26.8
python-dateutil==2.8.2
python-dotenv==0.19.2
